+++ R/DCS_MainFunctions.R
@@ -4,6 +4,13 @@
 #                                                                              #
 ################################################################################
 
+# Includes the main functions for the DCS package
+
+# set.options (exported)
+
+# dcs (exported)
+
+
 #------------------------Set Options via Function------------------------------#
 
 #' Set Options for the DCS procedure
@@ -17,18 +24,18 @@
 #' @param drv A non-negative vector of length 2, containing the derivative
 #'  orders to be estimated from the given data. The default is \code{c(0, 0)}. For
 #'  LP-regression, polynomial order is selected as \eqn{(\nu_1 + 1, \nu_2 + 1)}.
-#' @param var_est the method of estimating the variance coefficient \eqn{c_f}. 
-#'  Currently available are \code{var_est = "iid"} (the default), where an iid. 
-#'  error structure is modeled and \code{var_est = "qarma"}, where a spatial
-#'  ARMA model is applied. Use \code{"qarma_gpac"} for automated order 
-#'  selection.
+#' @param var_model the method of estimating the variance coefficient \eqn{c_f}. 
+#'  Currently available are \code{var_model = c("iid", "sarma_HR", "sarma_sep",
+#'  "sarma_RSS", "sfarima_RSS")}. Replacing the argument \code{var_model}. For
+#'  code using \code{var_est}, the argument is converted to \code{var_model}.
 #' @param ... Additional arguments passed to \code{set.options()}. This includes
 #'  \code{IPI_options}, a list containing further options used by the iterative
 #'  plug-in algorithm.
 #' 
 #' @section Details:
 #' This function is used to set the options for bandwidth selection in the 
-#' \code{dcs} function. Detailed information can be found in the vignette.
+#' \code{dcs} function.
+#' Detailed information can be found in the vignette.
 #' 
 #' @return  An object of class \code{"dcs_options"}.
 #' 
@@ -41,7 +48,7 @@
 #' 
 #' set.options()
 #' 
-#' myOpt <- set.options(type = "KR", var_est = "qarma")
+#' myOpt <- set.options(type = "KR", var_model = "iid")
 #' y <- y.norm1 + matrix(rnorm(101^2), nrow = 101, ncol = 101)
 #' dcs(y, dcs_options = myOpt)
 
@@ -51,36 +58,43 @@ set.options <- function(    # inside function with default values in arguments
   # "KR" for kernel regression
   kerns     = c("MW_220", "MW_220"), # choose a kernel function
   drv       = c(0, 0),
-  var_est   = "iid",
+  var_model = "iid",
   
   # advanced options in ellipsis
   ...
 )
 {
   # get ellipsis
-  IPI_options <- list(...)
-  if (exists("IPI_options", IPI_options))
+  add_options <- list(...)
+  # include old var_est options
+  if (exists("var_est", add_options))
   {
-    IPI_options = IPI_options$IPI_options
+    message("Note: option \"var_est\" is deprecated, argument is converted to ",
+            "\"var_model\" automatically.")
+    if (add_options$var_est == "iid") { var_model = "iid" }
+    if (add_options$var_est == "qarma") { var_model = "sarma_HR" }
+    if (add_options$var_est == "sarma") { var_model = "sarma_sep" }
+    if (add_options$var_est == "sarma2") { var_model = "sarma_RSS" }
+    if (add_options$var_est == "lm") { var_model = "sfarima_RSS" }
+    if (add_options$var_est %in% c("qarma_gpac", "qarma_bic"))
+    {
+      add_options$var_model = "sarma_HR"
+      warning("For automatic order selection, use \"model_order\" in ",
+              "\"dcs()\".")
+    }  
+  }
+  if (exists("IPI_options", add_options))
+  {
+    IPI_options = add_options$IPI_options
+  } else {
+    IPI_options = list()
   }
   
   # check if inputs are vectors
   if (length(kerns) == 1) { kerns <- c(kerns, kerns) }
   if (length(drv) == 1) { drv <- c(drv, drv) }
   
-  exception.check.options.input(type, kerns, drv, var_est, IPI_options)
-  
-  # if (var_est == "lm" && type == "KR")
-  # {
-  #   warning("Long-Memory bandwidth selection only supported for local ", 
-  #           "polynomial regression. Type set to \"LP\".")
-  #   type <- "LP"
-  # }
-  if (var_est == "lm")
-  {
-    message("Estimation under long-memory errors (SFARIMA) is currently in ", 
-            "experimental state.")
-  }
+  exception.check.options.input(type, kerns, drv, var_model, IPI_options)
   
   # Select options according to type ("LP", "KR")
   if (type == "LP")
@@ -96,9 +110,6 @@ set.options <- function(    # inside function with default values in arguments
     {
       IPI_options$const_window <- FALSE
     }
-    
-    options_list <- list(type = type, kerns = kerns, p_order = p_order,
-                        drv = drv, var_est = var_est, IPI_options = IPI_options)
   } else if (type == "KR") {
     p_order <- NA
     if (!exists("infl_exp", IPI_options))
@@ -111,22 +122,23 @@ set.options <- function(    # inside function with default values in arguments
     {
       IPI_options$const_window <- FALSE
     }
-    
-    options_list <- list(type = type, kerns = kerns, p_order = p_order,
-                        drv = drv, var_est = var_est, IPI_options = IPI_options)
   } else {
     stop("Unknown type \"", type, "\"")
   }
   
   # change IPI_options for long memory estimation
-  if (!exists("infl_exp", IPI_options) && var_est == "lm")
+  if (!exists("infl_exp", IPI_options) && var_model == "sfarima_RSS")
   {
     IPI_options$infl_exp <- c(0.5, 0.5)
   }
-  if (!exists("infl_par", IPI_options) && var_est == "lm")
+  if (!exists("infl_par", IPI_options) && var_model == "sfarima_RSS")
   {
     IPI_options$infl_par <- c(3, 1)
   }
+  
+  options_list <- list(type = type, kerns = kerns, p_order = p_order,
+                       drv = drv, var_model = var_model,
+                       IPI_options = IPI_options)
   
   # apply class to output object
   class(options_list) <- "dcs_options"
@@ -154,13 +166,15 @@ set.options <- function(    # inside function with default values in arguments
 #'  carried out by the iterative plug-in algorithm.
 #' @param ... Additional arguments passed to \code{dcs}. These might include
 #'  numerical vectors \code{X} and/or \code{T} containing the exogenous
-#'  covariates with respect to the rows and columns. If \code{var_est} is any of
-#'  \code{"qarma"}, \code{"sarma"} or \code{"lm"}, \code{model_order} is an 
-#'  optional list containing the two-dimensional model orders in the form 
-#'  \code{list(ar = c(1, 1), ma = c(1, 1)}. If \code{var_est} is any of
-#'  \code{"qarma_gpac"} or \code{"qarma_bic"}, \code{order_max} an optional list
-#'  containing the two dimensional maximum orders for the order selection of the
-#'  QARMA estimation in the form \code{list(ar = c(1, 1), ma = c(1, 1)}.
+#'  covariates with respect to the rows and columns. If the error term model in 
+#'  \code{set.options()$var_model} is any of the available parametric models
+#'  (SARMA, SFARIMA), \code{model_order} is either an optional list containing
+#'  the two-dimensional model orders in the form
+#'  \code{list(ar = c(1, 1), ma = c(1, 1)} or a character string specifying
+#'  an order selection criterion by \code{c("AIC", "BIC", "gpac")}. In the case
+#'  of automatic order selection, \code{order_max} is an optional list
+#'  containing the two-dimensional maximum orders for the order selection in the
+#'  form \code{list(ar = c(1, 1), ma = c(1, 1)}.
 #' 
 #' @return \code{DCSmooth} returns an object of class "dcs", including
 #'  \tabular{ll}{
@@ -208,7 +222,7 @@ dcs <- function(Y, dcs_options = set.options(), h = "auto", ...)
   {
     # set up vectors for X and T if neccessary
     args_list <- list(...)
-    add_options <- list() # additional options depending on type and var_est
+    add_options <- list() # additional options depending on type and var_model
     
     if (!exists("X", where = args_list))
     {
@@ -224,34 +238,33 @@ dcs <- function(Y, dcs_options = set.options(), h = "auto", ...)
     }
     exception.check.XT(Y, X, T)
     
-    # qarma order if var_est == "qarma"
-    if (exists("model_order", where = args_list) &&
-        dcs_options$var_est %in% c("qarma", "lm", "sarma", "sarma2"))
+    # model order
+    if (exists("model_order", where = args_list))
     {
       exception.check.model_order(args_list$model_order, dcs_options)
-      add_options$model_order <- args_list$model_order
-    } else if (!exists("model_order", where = args_list) &&
-               dcs_options$var_est %in% c("qarma", "lm", "sarma", "sarma2")) {
-      add_options$model_order <- list(ar = c(1, 1), ma = c(1, 1))
+      add_options$model_order = args_list$model_order
+    } else if (dcs_options$var_model %in% dcs_list_var_model &&
+               dcs_options$var_model != "iid") {
+      add_options$model_order = list(ar = c(1, 1), ma = c(1, 1))
     }
-    # maximum order of order selection is applied
     if (exists("order_max", where = args_list) &&
-        ((dcs_options$var_est == "qarma_gpac") ||
-        (dcs_options$var_est == "qarma_bic")))
+        length(add_options$model_order) == 1 &&
+        !is.list(add_options$model_order) &&
+        add_options$model_order %in% c("aic", "bic", "gpac"))
     {
-      exception.check.model_order(args_list$order_max, dcs_options)
-      add_options$order_max <- args_list$order_max
-    } else if (!exists("order_max", where = args_list) &&
-               ((dcs_options$var_est == "qarma_gpac" ||
-               dcs_options$var_est == "qarma_bic"))) {
-      add_options$order_max <- list(ar = c(1, 1), ma = c(1, 1))
+      exception.check.order_max(args_list$order_max)
+      add_options$order_max = args_list$order_max
+    } else if (length(add_options$model_order) == 1 &&
+               !is.list(add_options$model_order) &&
+               add_options$model_order %in% c("aic", "bic", "gpac")) {
+      add_options$order_max = list(ar = c(1, 1), ma = c(1, 1))
     }
   }
 
   #-------Bandwidth Selection-------#
   
   # check if given bandwidths exist
-  if (h[1] == "auto")
+  if (length(h) == 1 && h[1] == "auto")
   {
     h_select_auto <- TRUE
     
@@ -273,8 +286,7 @@ dcs <- function(Y, dcs_options = set.options(), h = "auto", ...)
     
     if (h_select_obj$var_model$stnry == FALSE)
     {
-      warning("QARMA/SARMA model not stationary, a change of the option ",
-              "\"qarma_order\" is advised.")
+      warning("error model is not stationary.")
     }
     
     time_end <- Sys.time()
@@ -301,7 +313,7 @@ dcs <- function(Y, dcs_options = set.options(), h = "auto", ...)
                 "working value.")
       }
     
-    dcs_out <- KR_dcs_const0(yMat = Y, hVec = h_opt, drvVec = c(0, 0), 
+    Y_smth_out <- KR_dcs_const0(yMat = Y, hVec = h_opt, drvVec = c(0, 0), 
                             kernFcnPtrX = kernel_x,
                             kernFcnPtrT = kernel_t)
   } else if (dcs_options$type == "LP") {     # local polynomial regression
@@ -324,28 +336,28 @@ dcs <- function(Y, dcs_options = set.options(), h = "auto", ...)
                 "working value.")
       }
       
-    dcs_out <- LP_dcs_const0_BMod(yMat = Y, hVec = h_opt, polyOrderVec = 
+    Y_smth_out <- LP_dcs_const0_BMod(yMat = Y, hVec = h_opt, polyOrderVec = 
                             dcs_options$p_order, drvVec = dcs_options$drv,
                             muVec = mu_vec, weightFcnPtr_x = weight_x,
                             weightFcnPtr_t = weight_t)
   }
   
   # calculate residuals and estimate model
-  R <- Y - dcs_out
+  R <- Y - Y_smth_out
   var_model_est <- cf.estimation(R, dcs_options, add_options)
   
   if (h_select_auto == TRUE)
   {
-    dcs_out <- list(Y = Y, X = X, T = T, M = dcs_out, R = R, h = h_opt,
+    dcs_out <- list(Y = Y, X = X, T = T, M = Y_smth_out, R = R, h = h_opt,
                    c_f = h_select_obj$var_coef, 
-                   var_model = var_model_est$var_model,
+                   var_est = var_model_est$model_est,
                    dcs_options = dcs_options,
                    iterations = h_select_obj$iterations,
                    time_used = difftime(time_end, time_start, units = "secs"))
     attr(dcs_out, "h_select_auto") <- h_select_auto
   } else if (h_select_auto == FALSE) {
-    dcs_out <- list(X = X, T = T, Y = Y, M = dcs_out, R = R, h = h_opt,
-                   c_f = NA, var_model = var_model_est$var_model,
+    dcs_out <- list(X = X, T = T, Y = Y, M = Y_smth_out, R = R, h = h_opt,
+                   c_f = NA, var_est = var_model_est$model_est,
                    dcs_options = dcs_options,
                    iterations = NA, time_used = NA)
     attr(dcs_out, "h_select_auto") <- h_select_auto