+++ tests/testthat/test-dcs.R
@@ -52,23 +52,30 @@ context("Variance Estimation")
 
 test_that("model order is actually used", {
   set.seed(123)
+  ar = c(1, -0.3) %*% t(c(1, 0.2))
+  ma = c(1, 0.5) %*% t(c(1, -0.1))
   Y = y.norm1 + rnorm(101^2)
   model_order_test = list(ar = c(2, 1), ma = c(1, 3))
-  dcs_iid = dcs(Y, set.options(var_est = "iid"), model_order = model_order_test)
-  dcs_qarma = dcs(Y, set.options(var_est = "qarma"), 
+  dcs_iid = dcs(Y, set.options(var_model = "iid"),
+                model_order = model_order_test)
+  Y = y.norm1 + sarma.sim(101, 101, model = list(ar = ar, ma = ma, sigma = 1))$Y
+  dcs_qarma = dcs(Y, set.options(var_model = "sarma_HR"), 
                   model_order = model_order_test)
-  dcs_sarma = dcs(Y, set.options(var_est = "sarma"), 
+  dcs_sarma = dcs(Y, set.options(var_model = "sarma_sep"), 
                   model_order = model_order_test)
-  dcs_lm = dcs(Y, set.options(var_est = "lm"), model_order = model_order_test)
+  Y = y.norm1 + sfarima.sim(101, 101, model = list(ar = ar, ma = ma, 
+                                              d = c(0.1, 0.1), sigma = 1))$Y
+  dcs_lm = dcs(Y, set.options(var_model = "sfarima_RSS"),
+               model_order = model_order_test)
   
-  expect_equal(dim(dcs_iid$var_model$ar), NULL)
-  expect_equal(dim(dcs_iid$var_model$ma), NULL)
-  expect_equal(dim(dcs_qarma$var_model$ar), c(3, 2))
-  expect_equal(dim(dcs_qarma$var_model$ma), c(2, 4))
-  expect_equal(dim(dcs_sarma$var_model$ar), c(3, 2))
-  expect_equal(dim(dcs_sarma$var_model$ma), c(2, 4))
-  expect_equal(dim(dcs_lm$var_model$ar), c(3, 2))
-  expect_equal(dim(dcs_lm$var_model$ma), c(2, 4))
+  expect_equal(dim(dcs_iid$var_est$model$ar), NULL)
+  expect_equal(dim(dcs_iid$var_est$model$ma), NULL)
+  expect_equal(dim(dcs_qarma$var_est$model$ar), c(3, 2))
+  expect_equal(dim(dcs_qarma$var_est$model$ma), c(2, 4))
+  expect_equal(dim(dcs_sarma$var_est$model$ar), c(3, 2))
+  expect_equal(dim(dcs_sarma$var_est$model$ma), c(2, 4))
+  expect_equal(dim(dcs_lm$var_est$model$ar), c(3, 2))
+  expect_equal(dim(dcs_lm$var_est$model$ma), c(2, 4))
 })
 
 # Test if smoothing works correctly
@@ -92,35 +99,35 @@ test_that("error term model is estimated under given bandwidths", {
   Y = matrix(rnorm(101^2), nrow = 101, ncol = 101)
   dcs_KR = dcs(Y, h = c(0.1, 0.1), dcs_options = set.options(type = "KR"))
   dcs_LP = dcs(Y, h = c(0.1, 0.1), dcs_options = set.options(type = "LP"))
-  expect_equal(length(dcs_KR$var_model$sigma), 1)
-  expect_equal(is.numeric(dcs_KR$var_model$sigma), TRUE)
-  expect_equal(length(dcs_LP$var_model$sigma), 1)
-  expect_equal(is.numeric(dcs_LP$var_model$sigma), TRUE)
+  expect_equal(length(dcs_KR$var_est$model$sigma), 1)
+  expect_equal(is.numeric(dcs_KR$var_est$model$sigma), TRUE)
+  expect_equal(length(dcs_LP$var_est$model$sigma), 1)
+  expect_equal(is.numeric(dcs_LP$var_est$model$sigma), TRUE)
   
   ar_mat = matrix(c(1, 0.1, -0.3, -0.2, 0.1, 0.1), 3, 2)
   ma_mat = matrix(c(1, 0.1, 0.1, 0.1), 2, 2)
-  Y = y.norm1 + qarma.sim(101, 101, model = list(ar = ar_mat, ma = ma_mat, 
+  Y = y.norm1 + sarma.sim(101, 101, model = list(ar = ar_mat, ma = ma_mat, 
                                                  sigma = 1))$Y
   dcs_KR = dcs(Y, h = c(0.1, 0.1), dcs_options = set.options(type = "KR", 
-                                                             var_est = "qarma"),
+                                                      var_model = "sarma_HR"),
                model_order = list(ar = c(2, 1), ma = c(1, 1)))
   dcs_LP = dcs(Y, h = c(0.1, 0.1), dcs_options = set.options(type = "LP", 
-                                                             var_est = "qarma"),
+                                                      var_model = "sarma_HR"),
                model_order = list(ar = c(2, 1), ma = c(1, 1)))
-  expect_equal(length(dcs_KR$var_model$sigma), 1)
-  expect_equal(dim(dcs_KR$var_model$ar), c(3, 2))
-  expect_equal(dim(dcs_KR$var_model$ma), c(2, 2))
-  expect_equal(is.numeric(dcs_KR$var_model$sigma), TRUE)
-  expect_equal(is.numeric(dcs_KR$var_model$ar), TRUE)
-  expect_equal(is.numeric(dcs_KR$var_model$ma), TRUE)
-  expect_equal(dcs_KR$var_model$stnry, TRUE)
-  expect_equal(length(dcs_LP$var_model$sigma), 1)
-  expect_equal(dim(dcs_LP$var_model$ar), c(3, 2))
-  expect_equal(dim(dcs_LP$var_model$ma), c(2, 2))
-  expect_equal(is.numeric(dcs_LP$var_model$sigma), TRUE)
-  expect_equal(is.numeric(dcs_LP$var_model$ar), TRUE)
-  expect_equal(is.numeric(dcs_LP$var_model$ma), TRUE)
-  expect_equal(dcs_LP$var_model$stnry, TRUE)
+  expect_equal(length(dcs_KR$var_est$model$sigma), 1)
+  expect_equal(dim(dcs_KR$var_est$model$ar), c(3, 2))
+  expect_equal(dim(dcs_KR$var_est$model$ma), c(2, 2))
+  expect_equal(is.numeric(dcs_KR$var_est$model$sigma), TRUE)
+  expect_equal(is.numeric(dcs_KR$var_est$model$ar), TRUE)
+  expect_equal(is.numeric(dcs_KR$var_est$model$ma), TRUE)
+  expect_equal(dcs_KR$var_est$stnry, TRUE)
+  expect_equal(length(dcs_LP$var_est$model$sigma), 1)
+  expect_equal(dim(dcs_LP$var_est$model$ar), c(3, 2))
+  expect_equal(dim(dcs_LP$var_est$model$ma), c(2, 2))
+  expect_equal(is.numeric(dcs_LP$var_est$model$sigma), TRUE)
+  expect_equal(is.numeric(dcs_LP$var_est$model$ar), TRUE)
+  expect_equal(is.numeric(dcs_LP$var_est$model$ma), TRUE)
+  expect_equal(dcs_LP$var_est$stnry, TRUE)
 })
 
 test_that("(too) small bandwidths are handled correctly", {